split kv api
