+++ Cargo.toml
@@ -19,6 +19,13 @@ members = [
     "common/progress",
     "common/raft-store",
     "common/runtime",
+    "common/streams",
+    "common/exception",
+    "common/tracing",
+    "common/profiling",
+    "common/store-api",
+    "common/store-api/util",
+    "common/store-api/kv-api",
     "common/sled-store",
     "common/stoppable",
     "common/store-api",
