+++ common/management/Cargo.toml
@@ -12,7 +12,8 @@ edition = "2021"
 common-exception= {path = "../exception"}
 common-metatypes= {path = "../metatypes"}
 common-runtime= {path = "../runtime"}
-common-store-api= {path = "../store-api"}
+common-store-kv-api= {path = "../store-api/kv-api"}
+common-store-kv-api-vo= {path = "../store-api/kv-api-vo"}
 common-tracing= {path = "../tracing"}
 
 kvlocal = {path = "../../kvlocal" }
