+++ common/management/src/namespace/namespace_mgr_test.rs
@@ -22,11 +22,11 @@ use common_metatypes::KVMeta;
 use common_metatypes::KVValue;
 use common_metatypes::MatchSeq;
 use common_metatypes::SeqValue;
-use common_store_api::kv_apis::kv_api::MGetKVActionResult;
-use common_store_api::kv_apis::kv_api::PrefixListReply;
-use common_store_api::GetKVActionResult;
-use common_store_api::KVApi;
-use common_store_api::UpsertKVActionResult;
+use common_store_kv_api::KVApi;
+use common_store_kv_api_vo::GetKVActionResult;
+use common_store_kv_api_vo::MGetKVActionResult;
+use common_store_kv_api_vo::PrefixListReply;
+use common_store_kv_api_vo::UpsertKVActionResult;
 use mockall::predicate::*;
 use mockall::*;
