+++ common/management/src/user/user_mgr_test.rs
@@ -19,11 +19,11 @@ use async_trait::async_trait;
 use common_exception::ErrorCode;
 use common_metatypes::KVMeta;
 use common_metatypes::MatchSeq;
-use common_store_api::kv_apis::kv_api::MGetKVActionResult;
-use common_store_api::kv_apis::kv_api::PrefixListReply;
-use common_store_api::GetKVActionResult;
-use common_store_api::KVApi;
-use common_store_api::UpsertKVActionResult;
+use common_store_kv_api::KVApi;
+use common_store_kv_api_vo::GetKVActionResult;
+use common_store_kv_api_vo::MGetKVActionResult;
+use common_store_kv_api_vo::PrefixListReply;
+use common_store_kv_api_vo::UpsertKVActionResult;
 use mockall::predicate::*;
 use mockall::*;
 use sha2::Digest;
