+++ common/store-api-sdk/Cargo.toml
@@ -17,8 +17,10 @@ common-planners = {path = "../planners"}
 common-runtime = {path = "../runtime"}
 common-streams = {path = "../streams"}
 common-tracing = {path = "../tracing"}
-common-store-api= {path = "../store-api"}
+common-store-api = {path = "../store-api"}
+common-store-kv-api-vo = {path = "../store-api/kv-api-vo"}
 common-infallible = {path = "../infallible"}
+common-store-api-sdk-util = {path = "./util"}
 
 # Github dependencies
