+++ common/store-api/kv-api-vo/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "common-store-kv-api-vo"
+version = "0.1.0"
+authors = ["Databend Authors <opensource@datafuselabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+common-metatypes = {path= "../../metatypes"}
+
+serde = { version = "1.0", features = ["derive"] }
