+++ common/store-api/src/lib.rs
@@ -22,11 +22,6 @@ pub use data_block_apis::data_block_api::ReadPlanResult;
 pub use data_block_apis::data_block_api::StorageApi;
 pub use data_block_apis::data_block_api::Summary;
 pub use data_block_apis::data_block_api::TruncateTableResult;
-pub use kv_apis::kv_api::GetKVActionResult;
-pub use kv_apis::kv_api::KVApi;
-pub use kv_apis::kv_api::PrefixListReply;
-pub use kv_apis::kv_api::UpsertKVActionResult;
-pub use kv_apis::kv_api_sync::SyncKVApi;
 pub use meta_apis::meta_api::CommitTableReply;
 pub use meta_apis::meta_api::CreateDatabaseActionResult;
 pub use meta_apis::meta_api::CreateTableActionResult;
@@ -39,6 +34,6 @@ pub use meta_apis::meta_api::GetTableActionResult;
 pub use meta_apis::meta_api::MetaApi;
 
 pub mod data_block_apis;
-pub mod kv_apis;
 pub mod meta_apis;
-pub mod util;
+pub use common_store_api_util as util;
+pub use common_store_kv_api as kv_api;
