+++ common/store-api/util/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "common-store-api-util"
+version = "0.1.0"
+authors = ["Databend Authors <opensource@datafuselabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+common-runtime = { path = "../../runtime"}
+common-tracing = { path = "../../tracing"}
+
+
+
+lazy_static = "1.4.0"
