+++ kvlocal/Cargo.toml
@@ -16,6 +16,10 @@ common-exception = {path = "../common/exception"}
 common-metatypes = {path = "../common/metatypes"}
 common-raft-store = {path = "../common/raft-store"}
 common-runtime = {path = "../common/runtime"}
+common-tracing = {path = "../common/tracing"}
+common-store-kv-api = {path = "../common/store-api/kv-api"}
+common-store-kv-api-vo = {path = "../common/store-api/kv-api-vo"}
+common-store-api-util = {path = "../common/store-api/util"}
 common-sled-store = {path = "../common/sled-store"}
 common-store-api = {path = "../common/store-api"}
 common-tracing = {path = "../common/tracing"}
