+++ metasrv/src/executor/kv_handlers.rs
@@ -19,14 +19,14 @@ use common_metatypes::LogEntry;
 use common_metatypes::Operation;
 use common_raft_store::state_machine::AppliedState;
 use common_store_api_sdk::kv_api_impl::GetKVAction;
-use common_store_api_sdk::kv_api_impl::GetKVActionResult;
 use common_store_api_sdk::kv_api_impl::KVMetaAction;
 use common_store_api_sdk::kv_api_impl::MGetKVAction;
-use common_store_api_sdk::kv_api_impl::MGetKVActionResult;
-use common_store_api_sdk::kv_api_impl::PrefixListReply;
 use common_store_api_sdk::kv_api_impl::PrefixListReq;
 use common_store_api_sdk::kv_api_impl::UpsertKVAction;
-use common_store_api_sdk::kv_api_impl::UpsertKVActionResult;
+use common_store_kv_api_vo::GetKVActionResult;
+use common_store_kv_api_vo::MGetKVActionResult;
+use common_store_kv_api_vo::PrefixListReply;
+use common_store_kv_api_vo::UpsertKVActionResult;
 
 use crate::executor::action_handler::RequestHandler;
 use crate::executor::ActionHandler;
